test: group vsock host tests module
