+++ crates/vsock-host/src/tests/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod support;
+
+mod command;
+mod connection;
+mod file;
