+++ crates/vsock-host/src/tests/support.rs
